requirements.txt
jupyter
python-dotenv
anthropic
googlemaps